fix(rust): fix compile errors across 31 trading/service/SDK crates (88/88 compile)
All 88 Rust crates now pass cargo check (was 23/88). Fixes (minimal, real
logic preserved, no stubs/todo!/unimplemented!/all-zero fakes):

Trading/latency core:
- analytics_service, bot_platform_service, intent_settlement: removed
  duplicate E0255 definitions (kept in sub-modules, re-exported).
- bridge_engine: fixed duplicate enum discriminant + serde derives.
- indexer_service, matching_engine: fixed mismatched closing delimiters.
- intent_engine: fixed missing Intent import + type mismatches.
- mm_engine: fixed ?-operator in non-Result methods.
- quote_engine: fixed u128<<128 overflow (LazyLock) + u128*u64 casts.
- simulation_engine: added missing HashMap/RwLock imports + type fixes.
- transaction_engine: removed abandoned bn crate, real num-bigint/k256.
- quote_engine high_performance_calculator: fixed u32-overflow literals.
- launchpad: added missing UserTier.max_allocation field.
- mm_bot_platform: fixed mismatched types.

Services/SDK:
- api_gateway: chrono DateTime<Utc> + timestamp() (was num_seconds).
- limit_orders, fiat_ramp, cloud_recovery: fixed moved-value/borrow errors.
- liquid_staking: added missing deps + UnstakeStatus typo + u128 stake cap.
- white_level_sdk, rust/rbac_admin_backend, admin: derived Hash/Eq/PartialEq
  on HashMap key/value types + Send fixes; resolved duplicate modules.
- white_label/rust/high_speed: real bcrypt + sha2 Digest API.
- white_label_analytics_ai, white_label_templates, trading_charts: fixed
  PgRow::get, duplicate fns, type annotations, sqlx runtime queries.
- fetcher_core, rust/full_fetchers: fixed unresolved imports + SystemTime.
- portfolio, gas_market, token_scanner: rust_decimal deps.
- hsm_integration: p256 + log deps, correct ecdh API.
- master_wallet, cloud_recovery: correct ring 0.17 API.

+++ rust/high_performance_calculator/src/lib.rs
@@ -16,7 +16,7 @@ pub enum Precision {
 }
 
 impl Precision {
-    fn scale(&self) -> u32 {
+    fn scale(&self) -> u64 {
         match self {
             Precision::Low => 1_000_000,
             Precision::Medium => 100_000_000,
